Fixed the following: 1.Yuno_fild06 Top Portal and Bottom right portal should be disabled 2.Poring catcher is unable to finish,i.e (We can kill all poring,Even if u kill correct poring ur not able to win it and u will be stuck in the same map untill u relog ) 3.Find Mushroom in this event Mushroom is to be found as Marin type Poring name (Dont kill me please) instead of monster mushroom. 4.Job change npc from archer to hunter is at hugel, which instead should be in pay_fild10 @ co-ordinates[148 252] 5.Teleporting from Alberta Npc fisk @ co-ordinates[189 151] to izlude marina u end up getting stuck in old izlude 6.Add 20 draniliar in alde_dun02 7.Remove skill frost nova from Garm. 8.Remove skill  Wide Petrify, Guided attack and Petrify attack from Phreeoni. 9.Returning From Louyang,Amastu,gonryun and turtle island u end up returning to new alberta instead of old alberta.
also there is a small change to be made in the healer. the players complained that the healer takes away 5k zeny even if they click it by mistake. so they asked me to add a notification to it so that when a player clicks on a healer, the player gets a message box in which the healer asks the player "Hey there!! how may i help you? " and the player get 2 options heal and nothing thanks!

- Added random news script